requirements.txt
streamlit
pandas
python-dotenv
sqlalchemy
pymysql
openai
scikit-learn
azure-storage-blob